refactor: repackage and get datasource from yaml
