--- src/main/java/com/programmers/vouchermanagement/configuration/properties/Domains.java
@@ -1,4 +1,0 @@
-package com.programmers.vouchermanagement.configuration.properties;
-
-public record Domains(Customer customer, Voucher voucher) {
-}
