--- src/main/java/com/programmers/vouchermanagement/configuration/properties/Resources.java
@@ -1,4 +1,0 @@
-package com.programmers.vouchermanagement.configuration.properties;
-
-public record Resources(String path, String buildPath) {
-}
