--- src/main/java/com/programmers/vouchermanagement/configuration/properties/Voucher.java
@@ -1,4 +1,0 @@
-package com.programmers.vouchermanagement.configuration.properties;
-
-public record Voucher(String fileName) {
-}
