+++ src/main/java/com/programmers/vouchermanagement/configuration/properties/file/Domains.java
@@ -1,0 +1,4 @@
+package com.programmers.vouchermanagement.configuration.properties.file;
+
+public record Domains(Customer customer, Voucher voucher) {
+}
