+++ src/main/java/com/programmers/vouchermanagement/configuration/properties/file/Resources.java
@@ -1,0 +1,4 @@
+package com.programmers.vouchermanagement.configuration.properties.file;
+
+public record Resources(String path, String buildPath) {
+}
