+++ src/main/java/com/programmers/vouchermanagement/configuration/properties/file/Voucher.java
@@ -1,0 +1,4 @@
+package com.programmers.vouchermanagement.configuration.properties.file;
+
+public record Voucher(String fileName) {
+}
